+++ location/example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:location/location.dart';
+import 'package:location_example/enable_in_background.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'get_location.dart';
@@ -45,7 +46,7 @@ class _MyHomePageState extends State<MyHomePage> {
               children: <Widget>[
                 const Text('Created by Guillaume Bernos'),
                 InkWell(
-                  child: Text(
+                  child: const Text(
                     'https://github.com/Lyokone/flutterlocation',
                     style: TextStyle(
                       decoration: TextDecoration.underline,
@@ -77,7 +78,7 @@ class _MyHomePageState extends State<MyHomePage> {
         title: Text(widget.title),
         actions: <Widget>[
           IconButton(
-            icon: Icon(Icons.info_outline),
+            icon: const Icon(Icons.info_outline),
             onPressed: _showInfoDialog,
           )
         ],
@@ -92,7 +93,9 @@ class _MyHomePageState extends State<MyHomePage> {
             Divider(height: 32),
             GetLocationWidget(),
             Divider(height: 32),
-            ListenLocationWidget()
+            ListenLocationWidget(),
+            Divider(height: 32),
+            EnableInBackgroundWidget(),
           ],
         ),
       ),