+++ location/lib/location.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:location_platform_interface/location_platform_interface.dart';
 
 export 'package:location_platform_interface/location_platform_interface.dart'
@@ -32,6 +30,16 @@ class Location {
     );
   }
 
+  /// Checks if service is enabled in the background mode.
+  Future<bool> isBackgroundModeEnabled() {
+    return LocationPlatform.instance.isBackgroundModeEnabled();
+  }
+
+  /// Enables or disables service in the background mode.
+  Future<bool> enableBackgroundMode({bool enable}) {
+    return LocationPlatform.instance.enableBackgroundMode(enable: enable);
+  }
+
   /// Gets the current location of the user.
   ///
   /// Throws an error if the app has no permission to access location.
@@ -49,7 +57,7 @@ class Location {
     return LocationPlatform.instance.hasPermission();
   }
 
-  /// Checks if the app has permission to access location.
+  /// Requests permission to access location.
   ///
   /// If the result is [PermissionStatus.deniedForever], no dialog will be
   /// shown on [requestPermission].